+++ V4/Turnover/dao_run_2.py
@@ -25,14 +25,15 @@ def func(m=None, n=None, group_size=None, lr=None, turnover_rate=None,
     dao = DAO(m=m, n=n, reality=reality, lr=lr, group_size=group_size)
     for period in range(search_loop):
         # Turbulence
-        if (period + 1) % 100 == 0:
-            reality.change(reality_change_rate=0.1)
-            for team in dao.teams:
-                for individual in team.individuals:
-                    individual.payoff = reality.get_payoff(belief=individual.belief)
-            # Turnover
+        # if (period + 1) % 100 == 0:
+        #     reality.change(reality_change_rate=0.1)
+        #     for team in dao.teams:
+        #         for individual in team.individuals:
+        #             individual.payoff = reality.get_payoff(belief=individual.belief)
+        # Turnover
+        if turnover_rate != 0:
             dao.turnover(turnover_rate=turnover_rate)
-            dao.search(threshold_ratio=0.5)
+        dao.search(threshold_ratio=0.5)
     return_dict[loop] = [dao.performance_across_time, dao.consensus_performance_across_time,
                          dao.diversity_across_time, dao.variance_across_time]
     sema.release()
@@ -41,13 +42,13 @@ def func(m=None, n=None, group_size=None, lr=None, turnover_rate=None,
 if __name__ == '__main__':
     t0 = time.time()
     m = 90
-    turnover_rate_list = [0.05, 0.10, 0.15, 0.20, 0.25]
+    turnover_rate_list = [0.5, 0.6, 0.7, 0.8, 0.9]
     group_size = 7
     n = 350
     lr = 0.3
     repetition = 50
     concurrency = 50
-    search_loop = 499
+    search_loop = 300
     # DVs
     performance_across_para = []
     consensus_performance_across_para = []
@@ -117,32 +118,33 @@ if __name__ == '__main__':
         # variance_across_para_time.append(variance_across_time)
     delay = np.random.uniform(1, 6)
     time.sleep(delay)
-    file_index = 1
-    performance_file_name = f"dao_performance_across_turnover_{file_index}"
+    performance_file_name = "dao_performance_across_turnover_2"
 
-    while os.path.exists(performance_file_name):
-        file_index += 1
-        performance_file_name = f"dao_performance_across_turnover_{file_index}"
+    if os.path.exists(performance_file_name):
+        with open("dao_performance_across_turnover_2", 'rb') as infile:
+            prior_performance = pickle.load(infile)
+        with open("dao_diversity_across_turnover_2", 'rb') as infile:
+            prior_diversity = pickle.load(infile)
+        with open("dao_variance_across_turnover_2", 'rb') as infile:
+            prior_variance = pickle.load(infile)
+        performance_final = [(each_1 + each_2) / 2 for each_1, each_2 in
+                             zip(prior_performance, performance_across_para)]
+        diversity_final = [(each_1 + each_2) / 2 for each_1, each_2 in zip(prior_diversity, diversity_across_para)]
+        variance_final = [(each_1 + each_2) / 2 for each_1, each_2 in zip(prior_variance, variance_across_para)]
+        with open("dao_performance_across_turnover_2", 'wb') as out_file:
+            pickle.dump(performance_final, out_file)
+        with open("dao_diversity_across_turnover_2", 'wb') as out_file:
+            pickle.dump(diversity_final, out_file)
+        with open("dao_variance_across_turnover_2", 'wb') as out_file:
+            pickle.dump(variance_final, out_file)
 
-    # save the without-time data (ready for figure)
-    with open("dao_performance_across_turnover_{0}".format(file_index), 'wb') as out_file:
-        pickle.dump(performance_across_para, out_file)
-    with open("consensus_performance_across_turnover_{0}".format(file_index), 'wb') as out_file:
-        pickle.dump(consensus_performance_across_para, out_file)
-    with open("dao_diversity_across_turnover_{0}".format(file_index), 'wb') as out_file:
-        pickle.dump(diversity_across_para, out_file)
-    with open("dao_variance_across_turnover_{0}".format(file_index), 'wb') as out_file:
-        pickle.dump(variance_across_para, out_file)
-
-    # save the with-time data
-    # with open("dao_performance_across_turnover_time", 'wb') as out_file:
-    #     pickle.dump(performance_across_para_time, out_file)
-    # with open("consensus_performance_across_turnover_time", 'wb') as out_file:
-    #     pickle.dump(consensus_performance_across_para_time, out_file)
-    # with open("dao_diversity_across_turnover_time", 'wb') as out_file:
-    #     pickle.dump(diversity_across_para_time, out_file)
-    # with open("dao_variance_across_turnover_time", 'wb') as out_file:
-    #     pickle.dump(variance_across_para_time, out_file)
+    else:
+        with open("dao_performance_across_turnover_2", 'wb') as out_file:
+            pickle.dump(performance_across_para, out_file)
+        with open("dao_diversity_across_turnover_2", 'wb') as out_file:
+            pickle.dump(diversity_across_para, out_file)
+        with open("dao_variance_across_turnover_2", 'wb') as out_file:
+            pickle.dump(variance_across_para, out_file)
 
     t1 = time.time()
     print(time.strftime("%H:%M:%S", time.gmtime(t1 - t0)))
