+++ V4/Turnover/hierarchy_run_2.py
@@ -25,19 +25,20 @@ def func(m=None, n=None, group_size=None, lr=None, turnover_rate=None,
     hierarchy = Hierarchy(m=m, n=n, reality=reality, lr=lr, group_size=group_size, p1=0.1, p2=0.9)
     for period in range(search_loop):
         # Turbulence
-        if (period + 1) % 100 == 0:
-            reality.change(reality_change_rate=0.1)
-            # update the individual payoff
-            for team in hierarchy.teams:
-                for individual in team.individuals:
-                    individual.payoff = reality.get_payoff(belief=individual.belief)
-            # update the manager payoff
-            for manager in hierarchy.superior.managers:
-                manager.payoff = reality.get_policy_payoff(policy=manager.policy)
-            # update the code payoff
-            hierarchy.superior.code_payoff = reality.get_policy_payoff(policy=hierarchy.superior.code)
+        # if (period + 1) % 100 == 0:
+        #     reality.change(reality_change_rate=0.1)
+        #     # update the individual payoff
+        #     for team in hierarchy.teams:
+        #         for individual in team.individuals:
+        #             individual.payoff = reality.get_payoff(belief=individual.belief)
+        #     # update the manager payoff
+        #     for manager in hierarchy.superior.managers:
+        #         manager.payoff = reality.get_policy_payoff(policy=manager.policy)
+        #     # update the code payoff
+        #     hierarchy.superior.code_payoff = reality.get_policy_payoff(policy=hierarchy.superior.code)
         # Turnover
-        hierarchy.turnover(turnover_rate=turnover_rate)
+        if turnover_rate != 0:
+            hierarchy.turnover(turnover_rate=turnover_rate)
         hierarchy.search()
     return_dict[loop] = [hierarchy.performance_across_time, hierarchy.superior.performance_average_across_time,
                          hierarchy.diversity_across_time, hierarchy.variance_across_time]
@@ -47,13 +48,13 @@ def func(m=None, n=None, group_size=None, lr=None, turnover_rate=None,
 if __name__ == '__main__':
     t0 = time.time()
     m = 90
-    turnover_rate_list = [0.05, 0.10, 0.15, 0.20, 0.25]
+    turnover_rate_list = [0.5, 0.6, 0.7, 0.8, 0.9]
     group_size = 7
     n = 350
     lr = 0.3
     repetition = 50
     concurrency = 50
-    search_loop = 499
+    search_loop = 300
     # DVs
     performance_across_para = []
     superior_performance_across_para = []
@@ -125,32 +126,33 @@ if __name__ == '__main__':
 
     delay = np.random.uniform(1, 6)
     time.sleep(delay)
-    file_index = 1
-    performance_file_name = f"hierarchy_performance_across_turnover_{file_index}"
+    performance_file_name = "hierarchy_performance_across_turnover_2"
 
-    while os.path.exists(performance_file_name):
-        file_index += 1
-        performance_file_name = f"hierarchy_performance_across_turnover_{file_index}"
+    if os.path.exists(performance_file_name):
+        with open("hierarchy_performance_across_turnover_2", 'rb') as infile:
+            prior_performance = pickle.load(infile)
+        with open("hierarchy_diversity_across_turnover_2", 'rb') as infile:
+            prior_diversity = pickle.load(infile)
+        with open("hierarchy_variance_across_turnover_2", 'rb') as infile:
+            prior_variance = pickle.load(infile)
+        performance_final = [(each_1 + each_2) / 2 for each_1, each_2 in
+                             zip(prior_performance, performance_across_para)]
+        diversity_final = [(each_1 + each_2) / 2 for each_1, each_2 in zip(prior_diversity, diversity_across_para)]
+        variance_final = [(each_1 + each_2) / 2 for each_1, each_2 in zip(prior_variance, variance_across_para)]
+        with open("hierarchy_performance_across_turnover_2", 'wb') as out_file:
+            pickle.dump(performance_final, out_file)
+        with open("hierarchy_diversity_across_turnover_2", 'wb') as out_file:
+            pickle.dump(diversity_final, out_file)
+        with open("hierarchy_variance_across_turnover_2", 'wb') as out_file:
+            pickle.dump(variance_final, out_file)
 
-    # save the without-time data
-    with open("hierarchy_performance_across_turnover_{0}".format(file_index), 'wb') as out_file:
-        pickle.dump(performance_across_para, out_file)
-    with open("superior_performance_across_turnover_{0}".format(file_index), 'wb') as out_file:
-        pickle.dump(superior_performance_across_para, out_file)
-    with open("hierarchy_diversity_across_turnover_{0}".format(file_index), 'wb') as out_file:
-        pickle.dump(diversity_across_para, out_file)
-    with open("hierarchy_variance_across_turnover_{0}".format(file_index), 'wb') as out_file:
-        pickle.dump(variance_across_para, out_file)
-
-    # save the with-time data
-    # with open("hierarchy_performance_across_time_TTT", 'wb') as out_file:
-    #     pickle.dump(performance_across_para_time, out_file)
-    # with open("superior_performance_across_time_TTT", 'wb') as out_file:
-    #     pickle.dump(superior_performance_across_para_time, out_file)
-    # with open("hierarchy_diversity_across_time_TTT", 'wb') as out_file:
-    #     pickle.dump(diversity_across_para_time, out_file)
-    # with open("hierarchy_variance_across_time_TTT", 'wb') as out_file:
-    #     pickle.dump(variance_across_para_time, out_file)
+    else:
+        with open("hierarchy_performance_across_turnover_2", 'wb') as out_file:
+            pickle.dump(performance_across_para, out_file)
+        with open("hierarchy_diversity_across_turnover_2", 'wb') as out_file:
+            pickle.dump(diversity_across_para, out_file)
+        with open("hierarchy_variance_across_turnover_2", 'wb') as out_file:
+            pickle.dump(variance_across_para, out_file)
 
     t1 = time.time()
     print(time.strftime("%H:%M:%S", time.gmtime(t1 - t0)))
